refactor(lowering): split declarations.rs into decl/ sub-module
