+++ crates/js-lowering/src/lower/decl/mod.rs
@@ -1,0 +1,5 @@
+mod class;
+mod function;
+mod import_export;
+mod pattern;
+mod variable;
